PSO - moved buffer-related components into a separate module
