+++ src/render/src/pso/mod.rs
@@ -16,14 +16,14 @@
 
 #![allow(missing_docs)]
 
+pub mod buffer;
 pub mod target;
 
 use std::default::Default;
 use std::marker::PhantomData;
 use gfx_core as d;
 use gfx_core::factory::Phantom;
-pub use gfx_core::pso::{Element, ElemOffset, ElemStride, Descriptor};
-use shade::ToUniform;
+pub use gfx_core::pso::{Descriptor};
 
 pub struct RawDataSet<R: d::Resources>{
     pub vertex_buffers: d::pso::VertexBufferSet<R>,
@@ -133,21 +133,7 @@ pub trait DataBind<R: d::Resources> {
     fn bind_to(&self, &mut RawDataSet<R>, &Self::Data, &mut d::handle::Manager<R>);
 }
 
-pub trait Structure<F> {
-    fn query(&str) -> Option<Element<F>>;
-}
 
-type AttributeSlotSet = usize;
-
-//Note: in order to increase the clarity of potential error messages,
-//it would make sense to put the generic bounds right here in the definitions.
-//However, this triggers Rust warning E0122, so will have to wait a bit.
-
-pub struct VertexBufferCommon<T, I>(AttributeSlotSet, PhantomData<(T, I)>);
-pub type VertexBuffer<T> = VertexBufferCommon<T, [(); 0]>;
-pub type InstanceBuffer<T> = VertexBufferCommon<T, [(); 1]>;
-pub struct ConstantBuffer<T: Structure<d::shade::ConstFormat>>(Option<d::ConstantBufferSlot>, PhantomData<T>);
-pub struct Global<T: ToUniform>(Option<d::shade::Location>, PhantomData<T>);
 pub struct ShaderResource<T>(Option<d::ResourceViewSlot>, PhantomData<T>);
 pub struct UnorderedAccess<T>(Option<d::UnorderedViewSlot>, PhantomData<T>);
 pub struct Sampler(Option<d::SamplerSlot>);
@@ -155,108 +141,6 @@ pub struct Sampler(Option<d::SamplerSlot>);
 /// It only makes sense for DX9 class hardware, since everything newer
 /// has samplers totally separated from the textures.
 pub struct TextureSampler<T>(ShaderResource<T>, Sampler);
-
-fn match_attribute(_: &d::shade::AttributeVar, _: d::format::Format) -> bool {
-    true //TODO
-}
-
-impl<'a,
-    T: Structure<d::format::Format>,
-    I: AsRef<[()]> + Default,
-> DataLink<'a> for VertexBufferCommon<T, I> {
-    type Init = ();
-    fn new() -> Self {
-        VertexBufferCommon(0, PhantomData)
-    }
-    fn is_active(&self) -> bool {
-        self.0 != 0
-    }
-    fn link_input(&mut self, at: &d::shade::AttributeVar, _: &Self::Init) ->
-                  Option<Result<d::pso::AttributeDesc, d::format::Format>> {
-        T::query(&at.name).map(|el| {
-            self.0 |= 1 << (at.slot as AttributeSlotSet);
-            if match_attribute(at, el.format) {
-                let rate = <I as Default>::default().as_ref().len();
-                Ok((el, rate as d::pso::InstanceRate))
-            }else {
-                Err(el.format)
-            }
-        })
-    }
-}
-
-impl<R: d::Resources, T, I> DataBind<R> for VertexBufferCommon<T, I> {
-    type Data = d::handle::Buffer<R, T>;
-    fn bind_to(&self, out: &mut RawDataSet<R>, data: &Self::Data, man: &mut d::handle::Manager<R>) {
-        let value = Some((man.ref_buffer(data.raw()).clone(), 0));
-        for i in 0 .. d::MAX_VERTEX_ATTRIBUTES {
-            if (self.0 & (1<<i)) != 0 {
-                out.vertex_buffers.0[i] = value;
-            }
-        }
-    }
-}
-
-impl<'a, T: Structure<d::shade::ConstFormat>>
-DataLink<'a> for ConstantBuffer<T> {
-    type Init = &'a str;
-    fn new() -> Self {
-        ConstantBuffer(None, PhantomData)
-    }
-    fn is_active(&self) -> bool {
-        self.0.is_some()
-    }
-    fn link_constant_buffer(&mut self, cb: &d::shade::ConstantBufferVar, init: &Self::Init) ->
-                            Option<Result<(), d::shade::ConstFormat>> {
-        if &cb.name == *init {
-            self.0 = Some(cb.slot);
-            Some(Ok(()))
-        }else {
-            None
-        }
-    }
-}
-
-impl<R: d::Resources, T: Structure<d::shade::ConstFormat>>
-DataBind<R> for ConstantBuffer<T> {
-    type Data = d::handle::Buffer<R, T>;
-    fn bind_to(&self, out: &mut RawDataSet<R>, data: &Self::Data, man: &mut d::handle::Manager<R>) {
-        if let Some(slot) = self.0 {
-            let value = Some(man.ref_buffer(data.raw()).clone());
-            out.constant_buffers.0[slot as usize] = value;
-        }
-    }
-}
-
-impl<'a, T: ToUniform> DataLink<'a> for Global<T> {
-    type Init = &'a str;
-    fn new() -> Self {
-        Global(None, PhantomData)
-    }
-    fn is_active(&self) -> bool {
-        self.0.is_some()
-    }
-    fn link_global_constant(&mut self, var: &d::shade::ConstVar, init: &Self::Init) ->
-                            Option<Result<(), d::shade::UniformValue>> {
-        if &var.name == *init {
-            //if match_constant(var, ())
-            self.0 = Some(var.location);
-            Some(Ok(()))
-        }else {
-            None
-        }
-    }
-}
-
-impl<R: d::Resources, T: ToUniform> DataBind<R> for Global<T> {
-    type Data = T;
-    fn bind_to(&self, out: &mut RawDataSet<R>, data: &Self::Data, _: &mut d::handle::Manager<R>) {
-        if let Some(loc) = self.0 {
-            let value = data.convert();
-            out.global_constants.push((loc, value));
-        }
-    }
-}
 
 impl<'a, T> DataLink<'a> for ShaderResource<T> {
     type Init = &'a str;
